renamed and added tests for unit check
