+++ services/web/server/tests/unit/isolated/test_catalog_handlers_utils.py
@@ -1,5 +1,5 @@
 from models_library.services import ServiceInput, ServiceOutput
-from simcore_service_webserver.catalog_handlers import can_connect
+from simcore_service_webserver.catalog_handlers_utils import can_connect
 
 
 def test_can_connect():
@@ -37,5 +37,42 @@ def test_can_connect():
     assert can_connect(
         from_output=ServiceOutput.parse_obj(input_sleeper_input_1),
         to_input=ServiceInput.parse_obj(file_picker_outfile),
+        strict=True,
+    )
+
+
+def test_can_connect_no_units_with_units():
+    port_without_unit = {
+        "label": "port W/O",
+        "description": "output w/o unit",
+        "type": "integer",
+    }
+
+    port_with_unit = {
+        "label": "port W/",
+        "description": "port w/ unit",
+        "type": "integer",
+        "unit": "m",  # <---
+    }
+
+    # w/o -> w
+    assert can_connect(
+        from_output=ServiceOutput.parse_obj(port_without_unit),
+        to_input=ServiceInput.parse_obj(port_with_unit),
+    )
+    assert not can_connect(
+        from_output=ServiceOutput.parse_obj(port_without_unit),
+        to_input=ServiceInput.parse_obj(port_with_unit),
+        strict=True,
+    )
+
+    # w -> w/o
+    assert can_connect(
+        from_output=ServiceOutput.parse_obj(port_with_unit),
+        to_input=ServiceInput.parse_obj(port_without_unit),
+    )
+    assert can_connect(
+        from_output=ServiceOutput.parse_obj(port_with_unit),
+        to_input=ServiceInput.parse_obj(port_without_unit),
         strict=True,
     )
